--- a/nim/config.nims
+++ b/nim/config.nims
@@ -9,8 +9,8 @@ const
 switch("nimcache", cacheSubdir)
 switch("opt", "speed")
 switch("panics", "off")
-switch("passC", "-s -flto")
-switch("passL", "-s -flto")
+switch("passC", "-flto=auto")
+switch("passL", "-flto=auto")
 
 when defined(profileGen):
   echo "Build with profileGen"
@@ -21,3 +21,4 @@ when defined(profileUse):
   echo "Build with profileUse"
   switch("passC", "-fprofile-use")
   switch("passL", "-fprofile-use")
+  switch("passL", "-s")