--- a/nim/config.nims
+++ b/nim/config.nims
@@ -1,18 +1,14 @@
 switch("opt", "speed")
-switch("cc", "gcc")
 switch("panics", "off")
 switch("passC", "-s -flto")
 switch("passL", "-s -flto")
 
 when defined(profileGen):
   echo "Build with profileGen"
-  switch("cc", "clang")
-  switch("passC", "-fprofile-instr-generate")
-  switch("passL", "-fprofile-instr-generate")
+  switch("passC", "-fprofile-generate")
+  switch("passL", "-fprofile-generate")
 
 when defined(profileUse):
   echo "Build with profileUse"
-  switch("cc", "clang")
-  switch("passC", "-fprofile-instr-use")
-  switch("passL", "-fprofile-instr-use")
-
+  switch("passC", "-fprofile-use")
+  switch("passL", "-fprofile-use")