--- a/nim/related.nimble
+++ b/nim/related.nimble
@@ -14,19 +14,16 @@ requires "nim >= 2.0.0"
 requires "jsony"
 requires "xxhash"
 
-const STAT_STEPS = 10
-
 task buildopt, "build optimized":
-  exec "gcc --version | head -1"
-  exec "clang --version | head -1"
-  exec "rm -f default.profdata default.profraw.*"
-  exec "nimble -d:profileGen -d:release build"
-  for i in 1..STAT_STEPS:
-    echo "Gen stat ", i, "/", STAT_STEPS
-    exec "./related > /dev/null"
-    mvFile "default.profraw", "default.profraw." & $i
-  when defined(macosx):
-    exec "xcrun llvm-profdata merge default.profraw.* --output default.profdata"
-  else:
-    exec "llvm-profdata merge default.profraw.* --output default.profdata"
-  exec "nimble -d:profileUse -d:release build"
+  exec "gcc --version 2>/dev/null | head -1"
+  exec "clang --version 2>/dev/null | head -1"
+  exec "rm -rf default*.prof* nimcache"
+  exec "nimble --silent build -d:profileGen -d:release"
+  echo "Gen stats"
+  exec "./related > /dev/null"
+  when defined(clang):
+    when defined(macosx):
+      exec "xcrun llvm-profdata merge default*.profraw --output default.profdata"
+    else:
+      exec "llvm-profdata merge default*.profraw --output default.profdata"
+  exec "nimble --verbose build -d:profileUse -d:release"