--- a/nim_con/.gitignore
+++ b/nim_con/.gitignore
@@ -1,3 +1,3 @@
 build/
-default.prof*
+default*.prof*
 nimcache/