--- a/nim_con/config.nims
+++ b/nim_con/config.nims
@@ -8,7 +8,6 @@ const
 
 switch("nimcache", cacheSubdir)
 
---cc:clang
 --mm:arc
 --outdir:build
 --tlsEmulation:off # default on|off varies by platform
@@ -16,12 +15,12 @@ switch("nimcache", cacheSubdir)
 
 when defined(profileGen):
   --hints:off
-  --passC:"-fprofile-instr-generate"
-  --passL:"-fprofile-instr-generate"
+  --passC:"-fprofile-generate"
+  --passL:"-fprofile-generate"
 
 when defined(profileUse):
-  --passC:"-fprofile-instr-use"
-  --passL:"-fprofile-instr-use"
+  --passC:"-fprofile-use"
+  --passL:"-fprofile-use"
 
 when defined(release):
   --passC:"-flto"