For `nim`, modify the PGO build steps. For `nim_con`, re-add the steps. In both cases the setup is for gcc as the default. If clang is the system compiler, e.g. on macOS, PGO will still be performed correctly.

gcc PGO seems to make a consistent difference. It also closes the gap a bit between `nim_con` and `nim`. Maybe that can be remedied with a multi-threaded SIMD impl for `nim_con` and/or a change in task lib, still hopeful for that!

It's outside the scope of this project, but noting again that with a higher number of cores the perf gains in `nim_con` (as is) relative to `nim` are much more significant, i.e it's not a waste of time.

Numbers below are from my Azure VM.

WITH *GCC* PGO
**Single Threaded**
`nim` on `nim_rev_5` branch
```
(related) michael@azure1:~/repos/related_post_gen$ ./nim.sh
  main
* nim_rev_5
Running Nim
...
Nim:

	Benchmark 1: ./related
	Processing time (w/o IO): 20.817ms
	Processing time (w/o IO): 20.767ms
	Processing time (w/o IO): 20.638ms
	Processing time (w/o IO): 20.795ms
	Processing time (w/o IO): 20.559ms
	Processing time (w/o IO): 20.703ms
	Processing time (w/o IO): 20.57ms
	Processing time (w/o IO): 20.591ms
	Processing time (w/o IO): 20.579ms
	Processing time (w/o IO): 20.559ms
	Processing time (w/o IO): 20.801ms
	Processing time (w/o IO): 20.513ms
	Processing time (w/o IO): 20.645ms
	 Time (mean ± σ):      39.2 ms ±   0.3 ms    [User: 29.9 ms, System: 9.3 ms]
	 Range (min … max):    38.7 ms …  39.7 ms    10 runs
	
Checking output
Success: related_posts_nim.json is valid!
Running Nim
Nim:

	Benchmark 1: ./related
	Processing time (w/o IO): 278.944ms
	Processing time (w/o IO): 279.605ms
	Processing time (w/o IO): 277.184ms
	 Time (mean ± σ):     354.7 ms ±   2.0 ms    [User: 318.8 ms, System: 35.8 ms]
	 Range (min … max):   353.2 ms … 356.1 ms    2 runs
	
Running Nim
Nim:

	Benchmark 1: ./related
	Processing time (w/o IO): 2537.252ms
	Processing time (w/o IO): 2524.897ms
	Processing time (w/o IO): 2527.73ms
	 Time (mean ± σ):      2.769 s ±  0.002 s    [User: 2.693 s, System: 0.072 s]
	 Range (min … max):    2.767 s …  2.770 s    2 runs
```
WITH *CLANG* PGO
**Single Threaded**
`nim` on `main` branch
```
(related) michael@azure1:~/repos/related_post_gen$ ./nim.sh 
* main
  nim_rev_5
Running Nim
...
Nim:

	Benchmark 1: ./related
	Processing time (w/o IO): 27.035ms
	Processing time (w/o IO): 27.042ms
	Processing time (w/o IO): 26.968ms
	Processing time (w/o IO): 26.949ms
	Processing time (w/o IO): 27.123ms
	Processing time (w/o IO): 26.935ms
	Processing time (w/o IO): 27.069ms
	Processing time (w/o IO): 27.136ms
	Processing time (w/o IO): 27.221ms
	Processing time (w/o IO): 27.105ms
	Processing time (w/o IO): 26.97ms
	Processing time (w/o IO): 26.879ms
	Processing time (w/o IO): 27.443ms
	 Time (mean ± σ):      50.4 ms ±   0.4 ms    [User: 40.2 ms, System: 10.2 ms]
	 Range (min … max):    49.7 ms …  51.1 ms    10 runs
	
Checking output
Success: related_posts_nim.json is valid!
Running Nim
Nim:

	Benchmark 1: ./related
	Processing time (w/o IO): 379.722ms
	Processing time (w/o IO): 378.044ms
	Processing time (w/o IO): 378.823ms
	 Time (mean ± σ):     471.8 ms ±   1.5 ms    [User: 435.8 ms, System: 35.9 ms]
	 Range (min … max):   470.7 ms … 472.8 ms    2 runs
	
Running Nim
Nim:

	Benchmark 1: ./related
	Processing time (w/o IO): 3354.719ms
	Processing time (w/o IO): 3347.557ms
	Processing time (w/o IO): 3355.413ms
	 Time (mean ± σ):      3.637 s ±  0.011 s    [User: 3.561 s, System: 0.070 s]
	 Range (min … max):    3.629 s …  3.644 s    2 runs
```
WITH *GCC* PGO
**Multi Threaded**
`nim_con` on `nim_rev_5` branch
```
(related) michael@azure1:~/repos/related_post_gen$ ./nim_con.sh
  main
* nim_rev_5
Running Nim Concurrent
using 4 threads
...
Nim Concurrent:

	Benchmark 1: ./build/related_con
	Processing time (w/o IO): 18.001ms
	Processing time (w/o IO): 17.105ms
	Processing time (w/o IO): 17.135ms
	Processing time (w/o IO): 17.848ms
	Processing time (w/o IO): 17.28ms
	Processing time (w/o IO): 17.144ms
	Processing time (w/o IO): 17.928ms
	Processing time (w/o IO): 17.156ms
	Processing time (w/o IO): 16.997ms
	Processing time (w/o IO): 16.95ms
	Processing time (w/o IO): 16.939ms
	Processing time (w/o IO): 17.826ms
	Processing time (w/o IO): 17.217ms
	 Time (mean ± σ):      38.1 ms ±   0.9 ms    [User: 56.1 ms, System: 8.0 ms]
	 Range (min … max):    37.0 ms …  39.4 ms    10 runs
	
Checking output
Success: related_posts_nim_con.json is valid!
Running Nim Concurrent
using 4 threads
Nim Concurrent:

	Benchmark 1: ./build/related_con
	Processing time (w/o IO): 225.389ms
	Processing time (w/o IO): 224.542ms
	Processing time (w/o IO): 226.59ms
	 Time (mean ± σ):     303.6 ms ±   1.4 ms    [User: 676.5 ms, System: 32.0 ms]
	 Range (min … max):   302.6 ms … 304.6 ms    2 runs
	
Running Nim Concurrent
using 4 threads
Nim Concurrent:

	Benchmark 1: ./build/related_con
	Processing time (w/o IO): 2309.263ms
	Processing time (w/o IO): 2398.246ms
	Processing time (w/o IO): 2282.576ms
	 Time (mean ± σ):      2.590 s ±  0.073 s    [User: 6.403 s, System: 0.072 s]
	 Range (min … max):    2.538 s …  2.642 s    2 runs
```
*WITHOUT* PGO
**Multi Threaded**
`nim_con` on `main` branch
```
(related) michael@azure1:~/repos/related_post_gen$ ./nim_con.sh 
* main
  nim_rev_5
Running Nim Concurrent
using 4 threads
...
Nim Concurrent:

	Benchmark 1: ./build/related_con
	Processing time (w/o IO): 19.388ms
	Processing time (w/o IO): 19.769ms
	Processing time (w/o IO): 19.159ms
	Processing time (w/o IO): 19.036ms
	Processing time (w/o IO): 19.037ms
	Processing time (w/o IO): 19.786ms
	Processing time (w/o IO): 19.263ms
	Processing time (w/o IO): 19.247ms
	Processing time (w/o IO): 19.235ms
	Processing time (w/o IO): 19.134ms
	Processing time (w/o IO): 19.51ms
	Processing time (w/o IO): 19.206ms
	Processing time (w/o IO): 19.231ms
	 Time (mean ± σ):      42.3 ms ±   0.7 ms    [User: 63.9 ms, System: 9.0 ms]
	 Range (min … max):    41.1 ms …  43.5 ms    10 runs
	
Checking output
Success: related_posts_nim_con.json is valid!
Running Nim Concurrent
using 4 threads
Nim Concurrent:

	Benchmark 1: ./build/related_con
	Processing time (w/o IO): 251.57ms
	Processing time (w/o IO): 254.711ms
	Processing time (w/o IO): 253.514ms
	 Time (mean ± σ):     341.1 ms ±   1.8 ms    [User: 757.5 ms, System: 34.0 ms]
	 Range (min … max):   339.9 ms … 342.4 ms    2 runs
	
Running Nim Concurrent
using 4 threads
Nim Concurrent:

	Benchmark 1: ./build/related_con
	Processing time (w/o IO): 2465.274ms
	Processing time (w/o IO): 2509.954ms
	Processing time (w/o IO): 2450.07ms
	 Time (mean ± σ):      2.755 s ±  0.040 s    [User: 6.967 s, System: 0.070 s]
	 Range (min … max):    2.727 s …  2.783 s    2 runs
```