@jinyus 2023-10-19
This PR:
```
Nim Concurrent:

        Benchmark 1: ./build/related_con
        Processing time (w/o IO): 28.026ms
        Processing time (w/o IO): 29.436ms
        Processing time (w/o IO): 27.904ms
        Processing time (w/o IO): 28.793ms
        Processing time (w/o IO): 27.425ms
        Processing time (w/o IO): 27.544ms
        Processing time (w/o IO): 28.665ms
        Processing time (w/o IO): 27.763ms
        Processing time (w/o IO): 27.525ms
        Processing time (w/o IO): 28.088ms
        Processing time (w/o IO): 27.792ms
        Processing time (w/o IO): 27.648ms
        Processing time (w/o IO): 27.81ms
          Time (mean ± σ):      51.8 ms ±   0.9 ms    [User: 89.9 ms, System: 9.0 ms]
          Range (min … max):    49.7 ms …  52.8 ms    10 runs
         
Checking output
Success: related_posts_nim_con.json is valid!
Generate 20000 posts
Run Benchmark (15k posts)
Running Nim Concurrent
using 4 threads
Nim Concurrent:

        Benchmark 1: ./build/related_con
        Processing time (w/o IO): 384.812ms
        Processing time (w/o IO): 386.282ms
        Processing time (w/o IO): 386.017ms
          Time (mean ± σ):     486.8 ms ±   0.4 ms    [User: 1126.1 ms, System: 35.4 ms]
          Range (min … max):   486.5 ms … 487.0 ms    2 runs
         
Generate 60000 posts
Run Benchmark (30k posts)
Running Nim Concurrent
using 4 threads
Nim Concurrent:

        Benchmark 1: ./build/related_con
        Processing time (w/o IO): 3390.228ms
        Processing time (w/o IO): 3302.047ms
        Processing time (w/o IO): 3376.324ms
          Time (mean ± σ):      3.679 s ±  0.046 s    [User: 9.294 s, System: 0.106 s]
          Range (min … max):    3.646 s …  3.711 s    2 runs


Nim:

        Benchmark 1: ./related
        Processing time (w/o IO): 41.159ms
        Processing time (w/o IO): 40.729ms
        Processing time (w/o IO): 41.047ms
        Processing time (w/o IO): 40.726ms
        Processing time (w/o IO): 40.704ms
        Processing time (w/o IO): 40.683ms
        Processing time (w/o IO): 41.027ms
        Processing time (w/o IO): 40.534ms
        Processing time (w/o IO): 40.691ms
        Processing time (w/o IO): 40.844ms
        Processing time (w/o IO): 41.02ms
        Processing time (w/o IO): 40.718ms
        Processing time (w/o IO): 40.661ms
          Time (mean ± σ):      65.9 ms ±   0.7 ms    [User: 55.1 ms, System: 10.7 ms]
          Range (min … max):    65.0 ms …  66.7 ms    10 runs
         
Nim:

        Benchmark 1: ./related
        Processing time (w/o IO): 606.42ms
        Processing time (w/o IO): 606.018ms
        Processing time (w/o IO): 606.251ms
          Time (mean ± σ):     706.5 ms ±   0.5 ms    [User: 669.0 ms, System: 37.3 ms]
          Range (min … max):   706.2 ms … 706.8 ms    2 runs
         
Nim:

        Benchmark 1: ./related
        Processing time (w/o IO): 5351.51ms
        Processing time (w/o IO): 5358.545ms
        Processing time (w/o IO): 5354.148ms
          Time (mean ± σ):      5.684 s ±  0.005 s    [User: 5.581 s, System: 0.093 s]
          Range (min … max):    5.680 s …  5.688 s    2 runs

```