@michaelsbradleyjr 2023-10-19
Wow, some big multipliers in your VM's numbers, I wonder if it's just really busy or something else is making a difference. The `5351.51ms` number of `nim` is 2x!

I did a fresh build/run of this PR branch in my Azure VM just now and am getting numbers that look very close to what I posted last night, see below.

Maybe try again later today?

```
Nim Concurrent:

        Benchmark 1: ./build/related_con
        Processing time (w/o IO): 17.163ms
        Processing time (w/o IO): 17.784ms
        Processing time (w/o IO): 17.121ms
        Processing time (w/o IO): 17.142ms
        Processing time (w/o IO): 17.272ms
        Processing time (w/o IO): 17.299ms
        Processing time (w/o IO): 17.386ms
        Processing time (w/o IO): 17.842ms
        Processing time (w/o IO): 17.135ms
        Processing time (w/o IO): 17.159ms
        Processing time (w/o IO): 17.164ms
        Processing time (w/o IO): 17.083ms
        Processing time (w/o IO): 17.267ms
          Time (mean ± σ):      39.2 ms ±   0.6 ms    [User: 55.4 ms, System: 9.9 ms]
          Range (min … max):    38.4 ms …  40.5 ms    10 runs
         
Nim Concurrent:

        Benchmark 1: ./build/related_con
        Processing time (w/o IO): 226.812ms
        Processing time (w/o IO): 229.003ms
        Processing time (w/o IO): 225.88ms
          Time (mean ± σ):     305.9 ms ±   2.3 ms    [User: 688.6 ms, System: 22.3 ms]
          Range (min … max):   304.3 ms … 307.5 ms    2 runs
         
Nim Concurrent:

        Benchmark 1: ./build/related_con
        Processing time (w/o IO): 2289.004ms
        Processing time (w/o IO): 2296.524ms
        Processing time (w/o IO): 2291.308ms
          Time (mean ± σ):      2.552 s ±  0.010 s    [User: 6.367 s, System: 0.092 s]
          Range (min … max):    2.544 s …  2.559 s    2 runs


Nim:

	Benchmark 1: ./related
	Processing time (w/o IO): 21.07ms
	Processing time (w/o IO): 20.774ms
	Processing time (w/o IO): 21.469ms
	Processing time (w/o IO): 20.914ms
	Processing time (w/o IO): 20.974ms
	Processing time (w/o IO): 20.806ms
	Processing time (w/o IO): 20.891ms
	Processing time (w/o IO): 20.759ms
	Processing time (w/o IO): 20.777ms
	Processing time (w/o IO): 20.766ms
	Processing time (w/o IO): 20.993ms
	Processing time (w/o IO): 21.097ms
	Processing time (w/o IO): 20.932ms
	 Time (mean ± σ):      42.3 ms ±   0.6 ms    [User: 34.0 ms, System: 8.3 ms]
	 Range (min … max):    41.1 ms …  43.4 ms    10 runs
	
Checking output
Success: related_posts_nim.json is valid!
Running Nim
Nim:

	Benchmark 1: ./related
	Processing time (w/o IO): 283.951ms
	Processing time (w/o IO): 282.81ms
	Processing time (w/o IO): 282.967ms
	 Time (mean ± σ):     369.7 ms ±   2.4 ms    [User: 339.6 ms, System: 29.9 ms]
	 Range (min … max):   368.0 ms … 371.3 ms    2 runs
	
Running Nim
Nim:

	Benchmark 1: ./related
	Processing time (w/o IO): 2541.093ms
	Processing time (w/o IO): 2541.75ms
	Processing time (w/o IO): 2525.562ms
	 Time (mean ± σ):      2.799 s ±  0.011 s    [User: 2.722 s, System: 0.072 s]
	 Range (min … max):    2.792 s …  2.807 s    2 runs
```