@jinyus 2023-10-19
```
gcc (GCC) 13.2.1 20230801                                                                                                                                                                                                                      
clang version 13.0.0 (https://github.com/apple/llvm-project.git 8ee3f51668ac68de50d541a815f00859f4922f98)                                                                                                                                      
Build with profileGen                                                                                                                                                                                                                          
lto-wrapper: warning: using serial compilation of 4 LTRANS jobs                                                                                                                                                                                
lto-wrapper: note: see the ‘-flto’ option documentation for more information
lto-wrapper: warning: using serial compilation of 2 LTRANS jobs
lto-wrapper: note: see the ‘-flto’ option documentation for more information
Hint: mm: orc; threads: on; opt: speed; options: -d:release
```