@jinyus 2023-10-19
I had a feeling it was compiler versions that's causing the issues. Very hard to solve with the current approach.

I followed this article to install docker, just a couple of commands: https://www.digitalocean.com/community/tutorials/how-to-install-and-use-docker-on-ubuntu-22-04